+++ src/rv_core/instruction/executor/d/fmv.rs
@@ -1,7 +1,7 @@
 use crate::rv_core::instruction::executor::prelude::*;
 
 pub fn xd(R { rd, rs1, rs2: _ }: R, x: &mut IntegerRegisters, f: &FloatRegisters) {
-    x[rd] = f[rs1].to_bits() as u64;
+    x[rd] = f[rs1].to_bits();
 }
 
 pub fn dx(R { rd, rs1, rs2: _ }: R, x: &IntegerRegisters, f: &mut FloatRegisters) {
